Correcciones menores de código
git-svn-id: http://svn-ctt.administracionelectronica.gob.es/svn/clienteafirma/trunk/project@640 dcdbb401-323c-49cf-88a0-803402807c85

+++ afirma-ui-simpleafirma/src/main/java/es/gob/afirma/standalone/LookAndFeelManager.java
@@ -46,7 +46,7 @@ public final class LookAndFeelManager {
         
         final  Object defaultFontHeight = Toolkit.getDefaultToolkit().getDesktopProperty("win.defaultGUI.font.height"); //$NON-NLS-1$
         if (defaultFontHeight != null && defaultFontHeight instanceof Integer) {
-           LARGE_FONT = ((Integer) defaultFontHeight) > LARGE_FONT_LIMIT;
+           LARGE_FONT = (((Integer) defaultFontHeight).intValue()) > LARGE_FONT_LIMIT;
         }
         // En Linux usmos siempre una configuracion como si se detectase un tamano de fuente grande 
         else if (Platform.OS.LINUX.equals(Platform.getOS())) {
